feat: add admin button in order to redirect to users page on click
