+++ src/app/app.component.ts
@@ -1,6 +1,6 @@
 import { NgFor, NgIf } from '@angular/common';
 import { Component } from '@angular/core';
-import { RouterOutlet } from '@angular/router';
+import { RouterOutlet, RouterLink } from '@angular/router';
 import { HeaderComponent } from './header/header.component';
 
 // Task 3
@@ -9,7 +9,7 @@ const newPaginations = [5, 4, 3, 2, 1];
 @Component({
   selector: 'app-root',
   standalone: true,
-  imports: [RouterOutlet, NgIf, NgFor, HeaderComponent],
+  imports: [RouterOutlet, NgIf, NgFor, HeaderComponent, RouterLink],
   templateUrl: './app.component.html',
   styleUrl: './app.component.scss',
 })
