+++ src/app/users-list/users-list.component.html
@@ -5,7 +5,7 @@
       <a href="#" class="user-email">{{user.email}}</a>
       <a href="#" class="user-website">{{user.website}}</a>
       <p class="user-company"><span class="user-company-bold">Company:</span> {{user.company.name}}</p>
-      <button class="delete-btn">Удалить пользователя</button>
+      <button class="delete-btn" (click)="deleteUser(user.id)">Удалить пользователя</button>
     </div>
   </div>
 </div>
