+++ src/app/users-list/users-list.component.ts
@@ -1,36 +1,7 @@
 import { NgFor } from '@angular/common';
 import { HttpClient } from '@angular/common/http';
 import { Component, inject } from '@angular/core';
-
-export interface IGeo {
-  lat: string;
-  lng: string;
-}
-
-export interface IAddress {
-  street: string;
-  suite: string;
-  city: string;
-  zipcode: string;
-  geo: IGeo;
-}
-
-export interface ICompany {
-  name: string;
-  catchPhrase: string;
-  bs: string;
-}
-
-export interface IUser {
-  id: number;
-  name: string;
-  username: string;
-  email: string;
-  address: IAddress;
-  phone: string;
-  website: string;
-  company: ICompany;
-}
+import { IUser } from './user.interface';
 
 @Component({
   selector: 'app-users-list',
@@ -48,5 +19,9 @@ export class UsersListComponent {
     this.apiService
       .get<IUser[]>('https://jsonplaceholder.typicode.com/users')
       .subscribe((response) => (this.users = response));
+  }
+
+  deleteUser(id: number) {
+    this.users = this.users.filter((user) => user.id !== id);
   }
 }
